Re-do file structure
