+++ src/main.js
@@ -1,6 +1,6 @@
 import { polyfill } from 'es6-object-assign'
 import { CalendarStyle } from './calendar-style.js'
-import DatePicker from './components/DatePicker.svelte'
+import DatePicker from './components/DatePickerAsyncWrapper.svelte'
 polyfill()
 
 export {
